Переводит статистику в один класс
