+++ src/utils/index.js
@@ -54,7 +54,7 @@ export const getTimeStr = (time1, time2) => {
   moment.locale(`en-gb`);
   const str1 = moment(time1).format(`LT`);
   const str2 = moment(time2).format(`LT`);
-  return `${str1} - ${str2}`;
+  return `${str1} — ${str2}`;
 };
 
 export const getRandomInRange = (min, max) => Math.floor(Math.random() * (max - min + 1) + min);
