+++ src/view/trip.js
